move listenOnUnused to a place where implementation code can use it
